+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Unreleased ###
---
+* Gravity Forms: Do not collect IP Address (GDPR compliance)
 
 ### 1.0.2: April 17th, 2018
 * Yoast SEO: breadcrumb class